Say "não consegui olhar" where the scan was going blind
Three collectors answered "nada aqui" for things they never managed to
read. It is the one failure mode this tool exists to prevent, and it was
in the middle of the most-used persistence surfaces.

cron never wrote PersistDenied["cron"] — the key existed, three checks
read it, and nothing ever filled it. The spool of the Debian is 1730
root:crontab: a scan without root fails to LIST it and reported zero
user crontabs, which is the same output as a host that has none.

sshd_config is 0600 on hardened hosts. Unreadable emptied SSH.Files, and
empty was the same state as "this machine has no SSH server" — the check
returned silent about the PermitRootLogin of a host that accepts root
over the network.

The SUID walk said "sem permissão neste galho: o resto da árvore
continua" in a comment and continued in silence. /home is 0700 per user
almost everywhere, and a setuid binary inside a home is retention that
survives the cleanup — precisely the branch that was skipped.

env.EhLacuna draws the line the three of them were missing: absence is an
ANSWER (half of the persistence paths do not exist on any given host),
permission denied is a GAP. Facts.listarNegando and Facts.lerNegando
apply it, so the distinction is made once instead of at every caller.

Two things the tests found on the way in:

  busybox does not use /etc/cron.daily or /var/spool/cron. Alpine puts
  jobs in /etc/crontabs and /etc/periodic/{15min,hourly,daily,weekly,
  monthly}, and Alpine is the base of most container images — scanning
  one returned zero schedules with the air of an answer.

  runPartsInterval had a default of 2592000, so every directory outside
  its four-name list came out "monthly". A job running every 15 minutes
  would appear as the slowest thing on the host instead of the fastest.
  Unknown cadence is now 0 — not determined, not monthly.

Also: the pcap filter classified ARP and LLDP as "não entendi", which
made a frame that simply is not IP indistinguishable from one that could
not be decoded, and the two test expectations that encoded that belief
are updated. And pcap_test did not compile on i686, where the socketcall
path and this writer had never been exercised.

+++ internal/pcap/pcap_test.go
@@ -97,7 +97,11 @@ func TestArquivoTemOFormatoDoLibpcap(t *testing.T) {
 
 	r := ler(t, buf.Bytes())
 	if r.magic != magicMicro {
-		t.Errorf("magic = %08x, queria %08x", r.magic, magicMicro)
+		// uint32 explícito: a constante não tipada vira `int` ao entrar num
+		// `any`, e 0xa1b2c3d4 estoura o int de 32 bits — o teste não compilava
+		// em i686, que é justamente a arquitetura onde o socketcall e este
+		// escritor nunca tinham sido exercitados.
+		t.Errorf("magic = %08x, queria %08x", r.magic, uint32(magicMicro))
 	}
 	if r.versao != [2]uint16{2, 4} {
 		t.Errorf("versão = %v, queria 2.4", r.versao)
